Major refactor of building search filters
- Refactor presenters to use new content builders
- Split search presenter class into several classes
- logic deferred to content builder on lots

* Tests on this will not pass - pushing to share code *
